Fix handling of partial packets
BufReader::fill_buf() only fills more data from the inner reader if the
internal buffer is empty. If a partial packet is read, calling
fill_buf() in process_packets_from() again wont read the rest of the
packet, resulting in an endless loop.

In addition, the PacketSize feature is parsed as a hex number by gdb, so
returning 65536 is actually interpreted as 0x65536.